+++ config.js
@@ -38,6 +38,7 @@ export default {
   displayPreview: true,
   displayOverview: true,
   displayOverviewsForChildren: false,
+  maxDisplayPixels: null,
   buildTileUrlTemplate: null,
   getMapSourceOptions: null,
   pathPrefix: "/",