+++ src/components/Catalog.vue
@@ -1,7 +1,7 @@
 <template>
   <b-card no-body :class="classes" v-visible.400="load" :img-placement="isList ? 'end' : undefined">
     <div class="card-img-wrapper">
-      <b-card-img v-if="hasImage" class="thumbnail" v-bind="thumbnail" lazy />
+      <AuthImage v-if="hasImage" class="thumbnail" v-bind="thumbnail" lazy />
     </div>
     <b-card-body>
       <b-card-title>
@@ -33,15 +33,16 @@ import StacLink from './StacLink.vue';
 import StacActions from './StacActions.vue';
 import { STAC } from 'stac-js';
 import { formatTemporalExtent } from '@radiantearth/stac-fields/formatters';
-import { BCard, BCardBody, BCardFooter, BCardImg, BCardText, BCardTitle } from 'bootstrap-vue-next';
+import { BCard, BCardBody, BCardFooter, BCardText, BCardTitle } from 'bootstrap-vue-next';
+import AuthImage from './AuthImage.vue';
 
 export default {
   name: 'Catalog',
   components: {
+    AuthImage,
     BCard,
     BCardBody,
     BCardFooter,
-    BCardImg,
     BCardText,
     BCardTitle,
     StacLink,