+++ src/components/HeaderTitle.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="header-title">
-    <img v-if="logo" :src="logo.getAbsoluteUrl()" :alt="logo.title" :title="logo.title" class="logo">
+    <AuthImage v-if="logo" :src="logo.getAbsoluteUrl()" :alt="logo.title" :title="logo.title" class="logo" />
     <span role="banner" :title="shortTitle">{{ shortTitle }}</span>
   </div>
 </template>
@@ -9,9 +9,13 @@
 import { mapState, mapGetters } from "vuex";
 import Utils from "../utils";
 import { getDisplayTitle } from "../models/stac";
+import AuthImage from "./AuthImage.vue";
 
 export default {
   name: "HeaderTitle",
+  components: {
+    AuthImage
+  },
   computed: {
     ...mapState(['catalogImage', 'catalogTitle', 'catalogTitleAfterImage', 'description', 'loading', 'url', 'uiLanguage']),
     ...mapGetters(['description', 'root', 'rootLink', 'title']),