+++ src/components/HrefActions.vue
@@ -47,9 +47,9 @@ import { mapGetters, mapState } from 'vuex';
 import AssetActions from '../../assetActions.config';
 import LinkActions from '../../linkActions.config';
 import AuthUtils from './auth/utils';
-import { Asset } from 'stac-js';
 import { browserProtocols } from 'stac-js/src/http';
 import { imageMediaTypes, geojsonMediaType, wozMediaTypes, zarrMediaTypes } from 'stac-js/src/mediatypes';
+import { getGeoZarrSourceOptionsFromAsset } from 'ol-stac/util.js';
 
 const disableDownloadTypes = [...zarrMediaTypes];
 const mapTypes = imageMediaTypes.concat([geojsonMediaType]).concat(wozMediaTypes);
@@ -100,7 +100,7 @@ export default {
   },
   computed: {
     ...mapState(['downloads', 'requestHeaders', 'buildTileUrlTemplate', 'useTileLayerAsFallback']),
-    ...mapGetters(['getRequestUrl']),
+    ...mapGetters(['getRequestUrl', 'isExternalUrl']),
     ...mapGetters('auth', ['isLoggedIn']),
     loading() {
       return Boolean(this.downloads[this.href]);
@@ -116,7 +116,18 @@ export default {
         return 'client';
       }
     },
+    // External services can't receive header-based credentials, so the
+    // external viewer actions won't work for data that requires them.
+    requiresHeaderCredentials() {
+      if (typeof this.data.href !== 'string') {
+        return false;
+      }
+      return size(this.requestHeaders) > 0 && !this.isExternalUrl(this.data.getAbsoluteUrl());
+    },
     actions() {
+      if (this.requiresAuth || this.requiresHeaderCredentials) {
+        return [];
+      }
       return Object.entries(this.isAsset ? AssetActions : LinkActions)
         .map(([id, plugin]) => new plugin(this.data, this, id))
         .filter(plugin => plugin.show);
@@ -137,6 +148,16 @@ export default {
       // Otherwise, all images that a browser can read are supported + GeoJSON
       else if (mapTypes.includes(this.data?.type)) {
         return true;
+      }
+      // Other Zarr assets can only be shown if the STAC metadata declares
+      // what to render (a datacube variable or bands)
+      else if (this.isAsset && zarrMediaTypes.includes(this.data?.type)) {
+        try {
+          const options = getGeoZarrSourceOptionsFromAsset(this.data, []);
+          return Boolean(options.variable || (options.bands && options.bands.length > 0));
+        } catch {
+          return false;
+        }
       }
       return false;
     },
@@ -303,12 +324,8 @@ export default {
       return '';
     },
     show() {
-      // Clone asset so that we can change the href.
-      // The this.href may include query parameters for authentication due to getRequestUrl.
-      // Unless we make authentication data available to ol-stac in a different way, we have to add it here.
-      const data = new Asset(this.data);
-      data.href = this.href;
-      this.$emit('show', data);
+      // Credentials are attached by the map itself (see MapMixin)
+      this.$emit('show', this.data);
     },
     handleAuthButton() {
       if (this.auth.length === 1) {