+++ src/components/Thumbnails.vue
@@ -4,8 +4,14 @@
       <FullscreenButton :element="() => $refs.previews" />
     </div>
     <div class="content">
-      <a v-for="thumbnail in thumbnails" :key="thumbnail.href" :href="thumbnail.getAbsoluteUrl()" target="_blank" rel="noopener noreferrer" download>
-        <img class="thumbnail" :src="thumbnail.getAbsoluteUrl()" :crossorigin="crossOriginMedia">
+      <a
+        v-for="thumbnail in thumbnails" :key="thumbnail.href" :href="hrefs[thumbnail.href] || thumbnail.getAbsoluteUrl()"
+        target="_blank" rel="noopener noreferrer" download
+      >
+        <AuthImage
+          class="thumbnail" :src="thumbnail.getAbsoluteUrl()" :crossorigin="crossOriginMedia"
+          @resolved="url => hrefs[thumbnail.href] = url"
+        />
       </a>
     </div>
   </div>
@@ -18,6 +24,7 @@ import { defineAsyncComponent } from 'vue';
 export default {
   name: 'Thumbnails',
   components: {
+    AuthImage: defineAsyncComponent(() => import('./AuthImage.vue')),
     FullscreenButton: defineAsyncComponent(() => import('./FullscreenButton.vue'))
   },
   props: {
@@ -25,6 +32,13 @@ export default {
       type: Array,
       required: true
     }
+  },
+  data() {
+    return {
+      // The URLs the thumbnails are shown from (incl. credentials / object URLs),
+      // so that the download links download what is shown
+      hrefs: {}
+    };
   },
   computed: {
     ...mapState(['crossOriginMedia'])